requirements.txt
fastapi
uvicorn
databases
sqlalchemy
joblib
scikit-learn
pandas
pymysql
python-multipart


google-cloud-bigquery
pandas-gbq